Completed all the capstone projects
Completed the finance capstone project as well that dealt with an analysis of the stocks of morgan stanley, chase, bank of america, citigroup, wells fargo and goldman sachs between the years of 2006 and 2016. It pulled the data from a pickle file and I plotted most of them using plotly. Only 1 uses pure matplotlib because I had some difficulty using plotly for plotting multiple traces/axes.